fix: add truncation detector composable and export it for re-use
